Final Refactions
This is a combination of 4 commits to change various items.

Refactored Edit Modal

Refactored Sale Modal

Refactored Edit Modal and Item + Title

Changed such that there is now a button to open the edit modal and the
delete button was moved there. Also added a title to the HTML page.

Refactor SearchBar